factor out utils from file_fetcher
